June 21, 2017 update, Build 239
Editable Max Fetch for Marsh -- Marsh horizontal erosion rates will now be applied when this editable marsh fetch has been exceeded for each marsh cell (previously this was hard-wired at 9km and this value will be loaded into older SLAMM6 simulations.).  This may be edited on a subsite by subsite basis.  This parameter is not relevant if the wave-energy erosion submodel is being utilized.

The Last year of simulation variable is always utilized even if run specific years is selected.  The interface was modified to make this more clear.